Refactor EloquentRepository for simplicity and fix few issues
