+++ .codeclimate.yml
@@ -14,6 +14,11 @@ engines:
     enabled: true
   phpmd:
     enabled: true
+    checks:
+      Design/TooManyPublicMethods:
+        enabled: false
+      Naming/ShortVariable:
+        enabled: false
 ratings:
   paths:
   - "**.inc"
