--- src/Traits/UsesContainer.php
@@ -1,54 +1,0 @@
-<?php
-
-/*
- * NOTICE OF LICENSE
- *
- * Part of the Rinvex Repository Package.
- *
- * This source file is subject to The MIT License (MIT)
- * that is bundled with this package in the LICENSE file.
- *
- * Package: Rinvex Repository Package
- * License: The MIT License (MIT)
- * Link:    https://rinvex.com
- */
-
-namespace Rinvex\Repository\Traits;
-
-use Illuminate\Contracts\Container\Container;
-
-trait UsesContainer
-{
-    /**
-     * The IoC container instance.
-     *
-     * @var \Illuminate\Contracts\Container\Container
-     */
-    protected $container;
-
-    /**
-     * Set the IoC container instance.
-     *
-     * @param \Illuminate\Contracts\Container\Container $container
-     *
-     * @return $this
-     */
-    public function setContainer(Container $container)
-    {
-        $this->container = $container;
-
-        return $this;
-    }
-
-    /**
-     * Return the IoC container instance or any of it's services.
-     *
-     * @param String $service
-     *
-     * @return mixed
-     */
-    public function getContainer($service = null)
-    {
-        return is_null($service) ? ($this->container ?: app()) : ($this->container[$service] ?: app($service));
-    }
-}
